+++ src/options/options.cpp
@@ -2,3 +2,4 @@
 
 ExperimentalOptions gExperimentalOptions;
 HighDpiOptions gHighDpiOptions;
+std::vector<KeyBindAction> gKeyBindings;
